ajustando responsividade

+++ src/components/WeatherInformations5Days/WeatherInformations5Days.jsx
@@ -15,7 +15,7 @@ function WeatherInformations5Days({ weather5Days }) {
     const next5DaysForecast = Object.values(dailyForecast).slice(1,6)
 
     function convertDate(date) {
-        const newDate = new Date(date.dt * 1000).toLocaleDateString('pt-BR', { weekday: 'long', day: '2-digit' })
+        const newDate = new Date(date.dt * 1000).toLocaleDateString('pt-BR', { weekday: 'short', day: '2-digit' })
         return newDate
     }
